Move VersionedManifest to db_state
